refactors query-builder into separate modules
updates build/test process to support modular query-builder
