+++ karma.conf.js
@@ -9,8 +9,16 @@ module.exports = function(config) {
     // base path, that will be used to resolve files and exclude
     basePath: '',
 
+    plugins: [
+      'karma-browserify',
+      'karma-jasmine',
+      'karma-phantomjs-launcher',
+      'karma-ng-json2js-preprocessor',
+      'karma-coverage'
+    ],
+
     // testing framework to use (jasmine/mocha/qunit/...)
-    frameworks: ['jasmine'],
+    frameworks: ['jasmine', 'browserify'],
 
     // list of files / patterns to load in the browser
     files: [
@@ -36,8 +44,13 @@ module.exports = function(config) {
    reporters: ['progress', 'coverage'],
 
    preprocessors: {
-      'src/*.js': ['coverage'],
+      'src/**/*.js': ['coverage'],
+      'src/browserify/*.js': ['browserify'],
       'test/**/*.json': ['ng-json2js']
+    },
+
+    browserify: {
+      transform: [ 'browserify-istanbul' ]
     },
 
     ngJson2JsPreprocessor: {
