+++ package.json
@@ -4,7 +4,10 @@
   "description": "An angular module that encapsulates the MarkLogic REST API builtins and normalizes the responses.",
   "dependencies": {},
   "devDependencies": {
+    "browserify": "^11.1.0",
+    "browserify-istanbul": "^0.2.1",
     "del": "^2.0.2",
+    "event-stream": "^3.3.1",
     "gulp": "^3.9.0",
     "gulp-concat": "^2.5.2",
     "gulp-gh-pages": "^0.5.2",
@@ -17,11 +20,14 @@
     "jsdoc": "^3.3.1",
     "jsdoc-baseline": "git://github.com/hegemonic/jsdoc-baseline.git#887a5a5",
     "karma": "^0.12.36",
-    "karma-coverage": "^0.4.2",
+    "karma-browserify": "^4.3.0",
+    "karma-coverage": "^0.5.2",
     "karma-jasmine": "^0.3.5",
     "karma-ng-json2js-preprocessor": "^1.0.0",
     "karma-phantomjs-launcher": "^0.2.0",
-    "phantomjs": "^1.9.17"
+    "phantomjs": "^1.9.17",
+    "vinyl-buffer": "^1.0.0",
+    "vinyl-source-stream": "^1.1.0"
   },
   "engines": {
     "node": ">=0.8.0"
